Store form configuration in a separate database table.
The saved form state was stored as a pickled dictionary with keys for each
form/specification.  This proved to be suboptimal for large applications with
many forms/specifications.  Now the state is stored in a database table with a
separate row for each user/form/specification/profile.  Apart from the
optimization, this will allow separate saving of configurations of form
profiles, but for now the profiles are unused.