Keep the current profile instance instead of its runtime construction
This commit is the start of replacing all the previous magic with filter list
management with a more straightforward approach.  The filter menu was
dynamically adding a new item or changing the current item automatically based
on detailed comparison of filtering conditions.  The original behavior seems
confusing when used with profiles, so profiles menu will not be updated
automatically.  Futher commits in this direction will follow.  This one makes
the current profile a static attribute instead of a function and replaces full
profile comparisons by comparing just their identifiers.