Get rid of sorting constant conversions in form vs. database sorting
Let's use the current conversion of saved user profiles for changing the
constants used for form sorting specification.  The form sorting constants are
now just aliases to database sorting constants which makes the mapping 1:1 and
all sorting conversions may be omited.  The profile conversion script will
replace the strings in saved user configurations, which was the only thing
preventing the change of the constant values for a long time.