Save the form size into the profile to be able to restore dualform sash position
The sash position is now not connected with the dualform itself, but rather
with its main form size.  As this size is now saved with the profile, the sash
will be initialized to the position given by the last saved size of the main
form's defualt profile.  It will currently not change when a profile is
changed, because there is no way for the main form to influence the dualform's
sash position.