Disable navigation in wx web forms
This change disables navigation anywhere outside the URL previously loaded by
the application.  Thus the web form can not be used as a web browser.  The
final goal would be to make the brower use proxy servers to allow finer control
over the URLs allowed to be browsed.  Until that works, we simply disable
browsing altogether.