Refactor custom toolbar control creation to be more versatile and less hacky
Some toolbar controls were created specifically by overriding the
add_toolbar_ctrl method, which was quite unclear and uncontrolled.  This is an
attempt to define a standard way to define such controls.  The control must be
implemented as a wx widget class and specified by a new `ctrl' argument of the
`UICommand' class.